Fix logic for filtering current day's metrics

+++ analysis/models.py
@@ -12,5 +12,5 @@ class DailyAnalysis(models.Model):
     )
     daily_avg_db_level = models.FloatField()
     daily_max_db_level = models.FloatField()
-    daily_no_of_exceedances = models.FloatField()
+    daily_no_of_exceedances = models.PositiveIntegerField()
     device = models.OneToOneField(Device, on_delete=models.CASCADE)
